==> preprocessing/00_wikitext_extraction.ipynb <==
import bz2
import xml.etree.ElementTree as ET
import json



input_file = "data/tawiki-latest-pages-articles.xml.bz2"
output_file = "data/tawiki_pages.jsonl"

NS = "{http://www.mediawiki.org/xml/export-0.11/}"

with bz2.open(input_file, "rb") as f, open(output_file, "w", encoding="utf-8") as out:
    context = ET.iterparse(f, events=("end",))

    for event, elem in context:
        if elem.tag == f"{NS}page":
            title_elem = elem.find(f"{NS}title")
            text_elem = elem.find(f".//{NS}text")

            title = title_elem.text if title_elem is not None else ""
            text = text_elem.text if text_elem is not None else ""

            if text:  # skip empty pages
                record = {
                    "title": title,
                    "wikitext": text
                }
                out.write(json.dumps(record, ensure_ascii=False) + "\n")

            # 🔥 VERY IMPORTANT: free memory
            elem.clear()
